refactor(database): remove SSL configuration from database setup
